third_party/blink/web_tests/platform/linux/virtual/webnn-service-with-gpu/external/wpt/webnn/conformance_tests/erf.https.any_gpu-expected.txt

This is a testharness.js-based test.
[FAIL] erf float16 0D scalar
  assert_array_approx_equals: test erf float16 property 0, expected -0.004909515380859375 +/- 0.001953125, expected -0.004909515380859375 but got 0
[FAIL] erf float16 1D constant tensor
  assert_array_approx_equals: test erf float16 property 0, expected -0.5947265625 +/- 0.001953125, expected -0.5947265625 but got 0
[FAIL] erf float16 1D tensor
  assert_array_approx_equals: test erf float16 property 0, expected -0.5947265625 +/- 0.001953125, expected -0.5947265625 but got 0
[FAIL] erf float16 2D tensor
  assert_array_approx_equals: test erf float16 property 0, expected -0.5947265625 +/- 0.001953125, expected -0.5947265625 but got 0
[FAIL] erf float16 3D tensor
  assert_array_approx_equals: test erf float16 property 0, expected -0.5947265625 +/- 0.001953125, expected -0.5947265625 but got 0
[FAIL] erf float16 4D tensor
  assert_array_approx_equals: test erf float16 property 0, expected -0.5947265625 +/- 0.001953125, expected -0.5947265625 but got 0
Harness: the test ran to completion.
